fix activation
